# docs/api-spec.yaml
openapi: '3.1.0'
info:
  title: Rabbit Maximizer API
  version: '0.1.0'
servers:
  - url: http://localhost:3000
    description: Local dev server

paths:
  /api/summary:
    get:
      operationId: getSummary
      parameters:
        - name: duration
          in: query
          schema:
            type: string
            enum: [24h, 2d, 3d, 5d, 1w]
            default: 24h
      responses:
        '200':
          description: Aggregate status summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Summary'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/queue:
    get:
      operationId: getQueue
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Paginated queue items
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedQueue'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/queue/order:
    get:
      operationId: getQueueOrder
      responses:
        '200':
          description: Pending items in effective order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueueOrderResponse'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/queue/order/move:
    post:
      operationId: moveQueueOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueueOrderMoveRequest'
      responses:
        '200':
          description: Updated effective order after move
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueueOrderResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Queue items not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/queue/order/move-to-top:
    post:
      operationId: moveToTop
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MoveToTopRequest'
      responses:
        '204':
          description: Item moved to top of queue
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Queue item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Item is not pending
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/events:
    get:
      operationId: getEvents
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        '200':
          description: Paginated events
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedEvents'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/config:
    get:
      operationId: getConfig
      responses:
        '200':
          description: Public client-facing configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicConfig'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/dashboard-state:
    get:
      operationId: getDashboardState
      parameters:
        - name: duration
          in: query
          schema:
            type: string
            enum: [24h, 2d, 3d, 5d, 1w]
            default: 24h
      responses:
        '200':
          description: Dashboard state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardState'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/queue/{uuid}/mark-reviewed:
    post:
      operationId: markReviewed
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Item marked reviewed
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok:
                    type: boolean
                additionalProperties: false
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Queue item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/queue/{uuid}/retrigger-now:
    post:
      operationId: retriggerNow
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
        - name: overridePause
          in: query
          required: false
          schema:
            type: boolean
      responses:
        '204':
          description: Retrigger posted
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Queue item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Scheduler is paused, item is not pending, or trigger failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/activity-list:
    get:
      operationId: getActivityList
      parameters:
        - name: since
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        '200':
          description: Paginated activity list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedActivityList'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/pause:
    post:
      operationId: setPaused
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PauseRequest'
      responses:
        '200':
          description: Pause state updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PauseResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/state/next_review_available_at:
    get:
      operationId: getNextReviewAvailable
      responses:
        '200':
          description: Next review availability timestamp
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StateResponse'
        '500':
          $ref: '#/components/responses/InternalError'

components:
  responses:
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
      additionalProperties: false

    QueueStatus:
      type: string
      enum: [pending, retriggered, resolved]

    QueueItem:
      type: object
      required:
        - id
        - uuid
        - repo_full_name
        - pr_number
        - pr_title
        - author_login
        - status
        - attempts
        - trigger_source
        - created_at
        - updated_at
      properties:
        id:
          type: integer
        uuid:
          type: string
        repo_full_name:
          type: string
        pr_number:
          type: integer
        pr_title:
          type: string
        author_login:
          type: string
          description: GitHub login of the PR author
        pr_state:
          type: string
          nullable: true
          description: PR state from GitHub (open, merged, closed)
          enum: [open, merged, closed]
        status:
          $ref: '#/components/schemas/QueueStatus'
        attempts:
          type: integer
        source_comment_url:
          type: string
        trigger_source:
          type: string
          enum: [dashboard_retrigger_now, scheduler]
        retrigger_comment_url:
          type: string
          nullable: true
        retriggered_at:
          type: string
          format: date-time
          nullable: true
        failed_at:
          type: string
          format: date-time
          nullable: true
        last_coderabbit_acknowledged_at:
          type: string
          format: date-time
          nullable: true
          description: When CodeRabbit last acknowledged a retrigger request for this PR (from pull_request table)
        reviewed_at:
          type: string
          format: date-time
          nullable: true
        resolved_at:
          type: string
          format: date-time
          nullable: true
          description: When the item transitioned to resolved (supersedes reviewed_at/failed_at for new code)
        resolution:
          type: string
          nullable: true
          description: Terminal reason when status='resolved' (review_completed, manual_review, pr_merged, pr_closed_without_merge, failed, skipped)
        coderabbit_review_state:
          type: string
          nullable: true
          description: Outcome of the most recent CodeRabbit review for this PR
          enum: [review_approved, review_changes_suggested]
        coderabbit_review_url:
          type: string
          nullable: true
          format: uri
          description: URL of the most recent CodeRabbit review for this PR
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      additionalProperties: false

    ActivityListItem:
      type: object
      required:
        - uuid
        - repo_full_name
        - pr_number
        - pr_title
        - author_login
        - status
        - created_at
        - source_comment_url
        - review_count
        - retrigger_count
        - last_activity_at
      properties:
        uuid:
          type: string
        repo_full_name:
          type: string
        pr_number:
          type: integer
        pr_title:
          type: string
        author_login:
          type: string
          description: GitHub login of the PR author
        pr_state:
          type: string
          nullable: true
          description: PR state from GitHub (open, merged, closed)
          enum: [open, merged, closed]
        status:
          $ref: '#/components/schemas/QueueStatus'
        resolution:
          type: string
          nullable: true
        retriggered_at:
          type: string
          format: date-time
          nullable: true
        resolved_at:
          type: string
          format: date-time
          nullable: true
        failed_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        retrigger_comment_url:
          type: string
          nullable: true
        source_comment_url:
          type: string
        last_review_url:
          type: string
          nullable: true
        last_review_state:
          type: string
          nullable: true
          enum: [review_approved, review_changes_suggested]
        review_count:
          type: integer
        retrigger_count:
          type: integer
        last_coderabbit_acknowledged_at:
          type: string
          format: date-time
          nullable: true
        last_activity_at:
          type: string
          format: date-time
      additionalProperties: false

    EventType:
      type: string
      enum:
        [
          coderabbit_review_approved,
          coderabbit_review_changes_suggested,
          coderabbit_review_skipped,
          coderabbit_run_id_changed,
          coderabbit_run_id_cleared,
          coderabbit_run_id_first_seen,
          detected,
          dismissed,
          enqueued,
          failed,
          retriggered,
        ]

    EventEntry:
      type: object
      required:
        - id
        - uuid
        - ts
        - type
        - repo_full_name
        - pr_number
        - correlation_id
        - version
        - payload
      properties:
        id:
          type: integer
        uuid:
          type: string
        ts:
          type: string
          format: date-time
        type:
          $ref: '#/components/schemas/EventType'
        repo_full_name:
          type: string
        pr_number:
          type: integer
        correlation_id:
          type: string
        request_id:
          type: string
        version:
          type: string
        metadata:
          type: object
          additionalProperties: true
        payload:
          type: object
          additionalProperties: true
      additionalProperties: false

    QueueCounts:
      type: object
      required: [pending, retriggered, resolved]
      properties:
        pending:
          type: integer
        retriggered:
          type: integer
        resolved:
          type: integer
      additionalProperties: false

    EventCounts:
      type: object
      required: [detected, enqueued, retriggered, failed]
      properties:
        detected:
          type: integer
        enqueued:
          type: integer
        retriggered:
          type: integer
        failed:
          type: integer
      additionalProperties: false

    DashboardState:
      type: object
      required: [lastSchedulerTickAt, nextReviewAvailableAt, pendingItems, skippedItems, eventCounts, paused, schedulerStale, trackedPrs]
      properties:
        lastSchedulerTickAt:
          type: string
          format: date-time
          nullable: true
        nextReviewAvailableAt:
          type: string
          format: date-time
          nullable: true
        pendingItems:
          type: array
          items:
            $ref: '#/components/schemas/QueueItem'
        skippedItems:
          type: array
          items:
            $ref: '#/components/schemas/QueueItem'
        eventCounts:
          $ref: '#/components/schemas/EventCounts'
        paused:
          type: boolean
        schedulerStale:
          type: boolean
        trackedPrs:
          type: array
          items:
            $ref: '#/components/schemas/TrackedPr'
      additionalProperties: false

    TrackedPr:
      type: object
      required: [repo_full_name, pr_number, title, author_login]
      properties:
        repo_full_name:
          type: string
        pr_number:
          type: integer
        title:
          type: string
        author_login:
          type: string
        last_review_state:
          type: string
          nullable: true
        last_coderabbit_review_at:
          type: string
          format: date-time
          nullable: true
      additionalProperties: false

    PauseRequest:
      type: object
      required: [paused]
      properties:
        paused:
          type: boolean
      additionalProperties: false

    PauseResponse:
      type: object
      required: [paused]
      properties:
        paused:
          type: boolean
      additionalProperties: false

    PublicConfig:
      type: object
      required: [pauseNotificationInitialDelaySec, pauseNotificationRepeatIntervalSec, schedulerStaleThresholdMs]
      properties:
        pauseNotificationInitialDelaySec:
          type: integer
          description: Delay in seconds before showing the first "paused" notification
        pauseNotificationRepeatIntervalSec:
          type: integer
          description: Interval in seconds between repeat "paused" notifications
        schedulerStaleThresholdMs:
          type: integer
          description: Time in ms before the scheduler is considered stale (no heartbeat)
      additionalProperties: false

    Summary:
      type: object
      required: [queueCounts, eventCounts, oldestPending]
      properties:
        queueCounts:
          $ref: '#/components/schemas/QueueCounts'
        eventCounts:
          $ref: '#/components/schemas/EventCounts'
        oldestPending:
          oneOf:
            - $ref: '#/components/schemas/QueueItem'
            - type: 'null'
      additionalProperties: false

    PaginatedQueue:
      type: object
      required: [data, total, page, pageSize]
      properties:
        data:
          type: array
          maxItems: 100
          items:
            $ref: '#/components/schemas/QueueItem'
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
      additionalProperties: false

    PaginatedActivityList:
      type: object
      required: [data, total, page, pageSize]
      properties:
        data:
          type: array
          maxItems: 100
          items:
            $ref: '#/components/schemas/ActivityListItem'
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
      additionalProperties: false

    PaginatedEvents:
      type: object
      required: [data, total, page, pageSize]
      properties:
        data:
          type: array
          maxItems: 100
          items:
            $ref: '#/components/schemas/EventEntry'
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
      additionalProperties: false

    MoveToTopRequest:
      type: object
      required: [queueItemUuid]
      properties:
        queueItemUuid:
          type: string
      additionalProperties: false

    QueueOrderMoveRequest:
      type: object
      required: [queueItemIds, direction]
      properties:
        queueItemIds:
          type: array
          minItems: 1
          maxItems: 100
          items:
            type: integer
        direction:
          type: string
          enum: [up, down]
      additionalProperties: false

    QueueOrderResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/QueueItem'
      additionalProperties: false

    StateResponse:
      type: object
      required: [next_review_available_at]
      properties:
        next_review_available_at:
          type: string
          format: date-time
          nullable: true
      additionalProperties: false